third_party/blink/web_tests/wpt_internal/event-timing/navigation/navigate-click-default.tentative.html

<!DOCTYPE html>
<html>
<meta charset=utf-8>
<title>Event Timing: navigate event from link click.</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=/event-timing/resources/event-timing-test-utils.js></script>
<body>
  <a id="navigateClick" href="#test">Click me</a>
<script>
  promise_test(async t => {
    assert_implements(window.PerformanceEventTiming, 'Event Timing is not supported');
    const isEnabled = window.internals && internals.runtimeFlags.navigationEventTimingEnabled;
    const entries = [];
    const observer = new PerformanceObserver((list) => {
      for (const entry of list.getEntries()) {
        entries.push(entry);
      }
    });
    observer.observe({ type: 'event', durationThreshold: 0 });

    const getEventCount = name => performance.eventCounts.get(name) || 0;
    assert_equals(performance.interactionCount, 0, 'Initial interactionCount should be 0');
    assert_equals(getEventCount('navigate'), 0, 'Initial navigate count should be 0');

    navigation.addEventListener('navigate', (event) => {
      mainThreadBusy(50);
    });

    const link = document.getElementById('navigateClick');
    link.href = '#test-' + Math.random();

    assert_false(entries.some(e => e.name === 'navigate'), 'Navigate entry should not be reported before click');

    const initialInteractionCount = performance.interactionCount;
    const initialNavigateCount = getEventCount('navigate');

    if (test_driver) {
      await test_driver.click(link);
    }

    const verifyMeasured = async () => {
      await t.step_wait(() => entries.some(e => e.name === 'click') &&
                             entries.some(e => e.name === 'navigate'),
                       'Wait for click and navigate entries');
      const clickId = entries.find(e => e.name === 'click').interactionId;
      const navigateEntries = entries.filter(e => e.name === 'navigate');
      assert_equals(navigateEntries.length, 1, 'Should have exactly one navigate entry');
      assert_greater_than(navigateEntries[0].interactionId, 0, 'Navigate entry should have a non-zero interactionId');
      assert_equals(navigateEntries[0].interactionId, clickId, 'Navigate entry should share interactionId with triggering event');

      assert_equals(performance.interactionCount, initialInteractionCount + 1, 'Interaction count should increase by 1');
      assert_equals(getEventCount('navigate'), initialNavigateCount + 1, 'Navigate event count should increase by 1');
    };

    const verifyNotMeasured = async () => {
      await new Promise(resolve => t.step_timeout(resolve, 500));
      assert_false(entries.some(e => e.name === 'navigate'), 'Navigate entry should NOT be reported');

      assert_equals(performance.interactionCount, initialInteractionCount + 1, 'Interaction count should increase by 1');
      assert_equals(getEventCount('navigate'), initialNavigateCount, 'Navigate event count should NOT increase');
    };

    if (isEnabled) {
      await verifyMeasured();
    } else {
      await verifyNotMeasured();
    }
  }, "Verify that navigate event from link click is measured.");
</script>
</body>
</html>
